+++ .github/workflows/main.yml
@@ -5,7 +5,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - name: Copy file to Ubuntu VM
+      - name: Copy alvaraalto.py to Ubuntu VM
         uses: appleboy/scp-action@master
         with:
           host: ${{ secrets.HOST }}
@@ -13,6 +13,15 @@ jobs:
           key: ${{ secrets.KEY }}
           source: "alvaraalto.py"
           target: "/home/opc/dc_bot"
+
+      - name: Copy src directory to Ubuntu VM
+        uses: appleboy/scp-action@master
+        with:
+          host: ${{ secrets.HOST }}
+          username: ${{ secrets.USERNAME }}
+          key: ${{ secrets.KEY }}
+          source: "src/*"
+          target: "/home/opc/dc_bot/src"
 
       - name: Restart Service
         uses: appleboy/ssh-action@v0.1.7
